DRA7XX - Late Attach - Update alloc func to align with no-map option
When performing late attach, it may be desireable to define the reserved
memory as "no-map" rather than as "reusable" in order to prevent access
by other kernel modules to the memory before the remoteproc driver has
a chance to allocate from the CMA pool and reclaim the memory. In the case
where the reserved memory is defined as "no-map", it becomes a DMA pool
rather than a CMA pool, and will not be mapped in the kernel.

When defining the memory as a DMA pool, the allocation logic changes. This
patch updates the allocation logic in the late attach code in u-boot to
align with the allocation logic for DMA pool memory in the kernel.

To use the DMA pool option, define the CONFIG_LATE_ATTACH_DMA_POOL flag
in the config file and also set the reserved mem pool for the late
attached cores to "no-map" in the kernel dtb file.

It is important to note that using a DMA pool may significantly change the
alignment and size of allocations due to the allocation scheme for DMA pools
in the kernel. The alignment requirements for a DMA pool allocation are more
strict than for a CMA pool, and therefore for the same remote core memory
requirements, more reserved memory may be needed.

Also note that if in the future the allocation logic changes in the kernel,
then this will also need to be updated.

Change-Id: I87f747fa946d8785295d6bd7f65622c641605cc2
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>