dra7xx: spl: add early boot functionality
This commit adds support for loading the DSP's and IPU's on the
dra7xx devices from MLO. The binaries to be loaded should be in
ELF format and placed in specific locations based the boot media.

SPL handles parsing the ELF binaries, bringing the corresponding
cores out of reset, copying the code/data sections to the DDR,
configure the MMU's and start the execution of the core.

Change-Id: I335e6a12cc7fb375fd13d083816e26555922a56a
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>