spl: sqpi: boot: Fix u-boot loading when SPL_OS_BOOT is not set
It's assumed that SPL should load the u-boot image when the
CONFIG_SPL_OS_BOOT option is not set.

The spl_spi_load_image() function uses the same variable to
check the returned value of function calls as well as to
determine if the u-boot image is to be loaded.  This variable
was initialized to 0, which isn't right because 0 means that
the u-boot image is not going to be loaded.

This problem caused HS boot to be broken.  GP boot was not
affected because CONFIG_SPL_OS_BOOT is defined non HS devices.

Change-Id: I00c3790c31618a8ab25cbb4f3cc848b5166f274a
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>