fastboot: erase QSPI boot areas only when necessary
When switching to QSPI in fastboot using "fastboot oem spi" command, the
default was to erase the MLO and u-boot.img regions. This requires us to
flash the binaries each time we use "fastboot oem spi" irrespective of
whether we want to flash the binaries.

1. Instead use "update" command to flash the binaries. This command erases
QSPI sectors only if necessary.

2. Remove the code to always erase QSPI in fastboot.

Change-Id: I59cac981373d23a9a1fcda28e0a0d96802b78197
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>