ti: dra76: Remove Vin1a_fld conflicting with VOUT1 HDMI power
For Dual HDMI use case, the VOUT1 port is connected to the onboard
HDMI chip which converts parallel video to HDMI.

For this to work, GPIO3_1 needs to be driven as the same is connected
for the power_on signal for this chip.

This pad conflicts with the VIN1A_FLD0 signal used for camera on the
JAMR board.
We do not need the FLD signal for vin1a for any JAMR use case
Hence remove this pinmux to solve the conflict and enable VOUT1 HDMI

Change-Id: Icea1a31dc2d8427a78b4ac563f649be420aef4cb
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>